Add files via upload
Magnetic connector separation now triggers timer (more than once too) and will allow GPS or Baro alt or timer to set cutter off, but only when LED is solid. What's left to do is to implement a way to set the "ground" alt for GPS or Baro and even swap between them based on lock/data reliability. I also need to implement the timer to go off regardless of whether the GPS or Baro alt set off the cutter.